chore: 회고탭 / 친구탭 기존 MVC 패턴 파일 제거
