+++ POME/Presentation/ViewControllers/Friend/FriendViewController.swift
@@ -1,93 +1,63 @@
 //
-//  FriendViewController.swift
+//  FriendTestViewController.swift
 //  POME
 //
-//  Created by 박지윤 on 2022/11/06.
+//  Created by 박소윤 on 2023/05/02.
 //
 
-import UIKit
-import Kingfisher
+import Foundation
+import RxSwift
+import RxCocoa
 
-class FriendViewController: BaseTabViewController, ControlIndexPath, Pageable {
+private extension IndexPath{
+    var ofFriendData: Int{
+        row - 1
+    }
+    var ofRecordData: Int{
+        row - 1
+    }
+}
 
-    var dataIndexBy: (IndexPath) -> Int = { indexPath in
-        return indexPath.row - 1
+final class FriendViewController: BaseTabViewController{
+    
+    private let viewModel: any FriendViewModelInterface
+    
+    init(viewModel: any FriendViewModelInterface = FriendViewModel()){
+        self.viewModel = viewModel
+        super.init(btnImage: Image.addPeople)
     }
     
-    //MARK: - Property
-    
-    var page: Int = 0
-    var isPaging: Bool = false
-    var hasNextPage: Bool = false
-    private var willLoadingViewAnimate: Bool{
-        page == 0 ? true : false
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
     }
     
-    var currentFriendIndex: Int = 0{
-        didSet{
-            page = 0
-            hasNextPage = false
-        }
+    private var isRefreshing: Bool{
+        guard let refreshControl = mainView.tableView.refreshControl else { return false }
+        return refreshControl.isRefreshing
     }
-    var currentEmotionSelectCardIndex: Int?
-
-    var friends = [FriendsResponseModel](){
-        didSet{
-            isTableViewEmpty()
-            if let friendsCell = friendView.tableView.cellForRow(at: [0,0], cellType: FriendListTableViewCell.self){
-                friendsCell.collectionView.reloadData()
-            }
-            friendImageManager.construct(by: friends)
-        }
-    }
-    private var willDelete = false
-    var records = [RecordResponseModel](){
-        didSet{
-            if(willDelete){
-                return
-            }
-            isPaging = false
-            isTableViewEmpty()
-            friendView.tableView.reloadData()
-        }
-    }
+    private var isLoading: Bool = false
     
-    let friendImageManager = FriendProfileImageManager.shared
-    let friendsChangeManager = FriendListChangeManager.shared
-    
-    //MARK: - UI
+    private let FRIEND_INFO_SECTION: Int = 0
+    private let COUNT_OF_NOT_RECORD_CELL: Int = 1
+    private let FRIEND_LIST_TABLEVIEW_CELL: IndexPath = [0,0]
     
     private typealias FriendListTableViewCell = FriendView.FriendListTableViewCell
     
-    let friendView = FriendView()
-    var emoijiFloatingView: ReactionFloatingView!
+    private let mainView: FriendView = FriendView()
+    private let reactionFloatingView: ReactionFloatingView = ReactionFloatingView()
     
-    //MARK: - Override
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        requestGetFriendsInitialize()
-    }
-    
-    override func viewWillAppear(_ animated: Bool) {
-        if(friendsChangeManager.isChange){
-            friendView.tableView.scrollToRow(at: [0,0], at: .top, animated: false)
-            requestGetFriendsInitialize()
-            friendsChangeManager.initialize()
-            return
-        }
-        requestGetFriends()
-        print(UserManager.token, UserManager.userId)
-    }
+    private let willRefreshData = BehaviorSubject<Void>(value: ())
+    private let willPaging = PublishSubject<Void>()
+    private let selectedFriendCellIndex = BehaviorRelay<Int>(value: 0)
     
     override func layout() {
         
         super.layout()
         
-        let tabBarHeight = self.tabBarController?.tabBar.frame.height
+        let tabBarHeight = tabBarController?.tabBar.frame.height
         
-        self.view.addSubview(friendView)
-        friendView.snp.makeConstraints{
+        view.addSubview(mainView)
+        mainView.snp.makeConstraints{
             $0.top.equalToSuperview().offset(Offset.VIEW_CONTROLLER_TOP)
             $0.leading.trailing.equalToSuperview()
             $0.bottom.equalToSuperview().offset(-tabBarHeight!)
@@ -95,364 +65,207 @@ class FriendViewController: BaseTabViewController, ControlIndexPath, Pageable {
     }
     
     override func initialize() {
-        friendView.tableView.delegate = self
-        friendView.tableView.dataSource = self
-        friendView.tableView.separatorStyle = .none
+        setTableViewDelegate()
+        setTableViewRefresh()
+    }
+    
+    private func setTableViewDelegate(){
+        mainView.tableView.delegate = self
+        mainView.tableView.dataSource = self
+        mainView.tableView.separatorStyle = .none
+    }
+    
+    private func setTableViewRefresh(){
+        let refreshControl = UIRefreshControl().then{
+            $0.backgroundColor = .white
+            $0.tintColor = .black
+            $0.addTarget(self, action: #selector(refreshingData), for: .valueChanged)
+        }
+        mainView.tableView.refreshControl = refreshControl
+    }
+    
+    @objc private func refreshingData(){
+        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
+            self.willRefreshData.onNext(())
+        }
+    }
+    
+    override func bind() {
+        
+        guard let viewModel = viewModel as? FriendViewModel else { return }
+        
+        viewModel.registerReactionCompleted = { [self] in
+            mainView.tableView.reloadRows(at: [[FRIEND_INFO_SECTION, $0 + COUNT_OF_NOT_RECORD_CELL]], with: .none)
+        }
+        
+        viewModel.hideRecordCompleted = { [self] in
+            mainView.tableView.deleteRows(at: [[FRIEND_INFO_SECTION, $0 + COUNT_OF_NOT_RECORD_CELL]], with: .fade)
+            ToastMessage.showHideCompleteMessage(in: self)
+        }
+        
+        let input = FriendViewModel.Input(
+            refreshView: willRefreshData.asObservable(),
+            willPaging: willPaging.asObservable(),
+            friendCellIndex: selectedFriendCellIndex.asObservable()
+        )
+        
+        let output = viewModel.transform(input)
+        
+        output.reloadFriends
+            .subscribe(onNext: { [self] in
+                mainView.tableView
+                    .cellForRow(at: FRIEND_LIST_TABLEVIEW_CELL, cellType: FriendListTableViewCell.self)?
+                    .collectionView.reloadData()
+            }).disposed(by: disposeBag)
+        
+        output.reloadRecords
+            .subscribe(onNext: { [self] in
+                if isLoading {
+                    isLoading = false
+                }
+                if isRefreshing {
+                    mainView.tableView.refreshControl?.endRefreshing()
+                }
+                isTableViewEmpty()
+                mainView.tableView.reloadData()
+            }).disposed(by: disposeBag)
+    }
+    
+    private func isTableViewEmpty(){
+        if viewModel.friends.isEmpty {
+            mainView.emptyViewWillShow(case: .nofriend)
+        } else if viewModel.records.isEmpty {
+            mainView.emptyViewWillShow(case: .noRecord)
+        } else {
+            mainView.emptyViewWillHide()
+        }
     }
     
     override func topBtnDidClicked() {
-        let vc = FriendSearchViewController()
-        vc.isFromRegister = false
-        self.navigationController?.pushViewController(vc, animated: true)
-    }
-    
-    //MARK: - Method
-    
-    private func isTableViewEmpty(){
-        if(friends.isEmpty){
-            friendView.emptyViewWillShow(case: .nofriend)
-        }else if(records.isEmpty){
-            friendView.emptyViewWillShow(case: .noRecord)
-        }else{
-            friendView.emptyViewWillHide()
+        let vc = FriendSearchViewController().then{
+            $0.isFromRegister = false
         }
+        navigationController?.pushViewController(vc, animated: true)
     }
 }
 
-extension FriendViewController{
-    
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        let offsetY = scrollView.contentOffset.y
-        let contentHeight = scrollView.contentSize.height
-        let height = scrollView.frame.height
-        
-        // 스크롤이 테이블 뷰 Offset의 끝에 가게 되면 다음 페이지를 호출
-        if offsetY > (contentHeight - height) {
-            if isPaging == false && hasNextPage {
-                beginPaging()
-            }
-        }
-    }
-    
-    func beginPaging(){
-        
-        isPaging = true
-        page = page + 1
-        
-        DispatchQueue.main.async { [self] in
-            friendView.tableView.reloadSections(IndexSet(integer: 1), with: .none)
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-            self.currentFriendIndex == 0 ? self.requestGetAllFriendsRecords() : self.requestGetFriendCards()
-        }
-    }
-}
-
-//MARK: - API
-extension FriendViewController{
-    
-    private func requestGetFriendsInitialize(){
-        currentFriendIndex = 0
-        FriendService.shared.getFriends(pageable: PageableModel(page: page)){ result in
-            switch result{
-            case .success(let data):
-                print("LOG: 'success' requestGetFriends", data)
-                self.friends = data
-                self.requestGetAllFriendsRecords()
-                break
-            default:
-                print("LOG: 'fail' requestGetFriends", result)
-                NetworkAlert.show(in: self){ [weak self] in
-                    self?.requestGetFriendsInitialize()
-                }
-                break
-            }
-        }
-    }
-    
-    private func requestGetFriends(){
-        FriendService.shared.getFriends(pageable: PageableModel(page: page)){ result in
-            switch result{
-            case .success(let data):
-                print("LOG: 'success' requestGetFriends", data)
-                self.friends = data
-                break
-            default:
-                print("LOG: 'fail' requestGetFriends", result)
-                NetworkAlert.show(in: self){ [weak self] in
-                    self?.requestGetFriends()
-                }
-                break
-            }
-        }
-    }
-    
-    private func requestGetAllFriendsRecords(){
-        FriendService.shared.getAllFriendsRecord(pageable: PageableModel(page: page),
-                                                 animate: willLoadingViewAnimate){ response in
-            switch response {
-            case .success(let data):
-                print("LOG: success requestGetAllFriendsRecords", data)
-                self.processResponseGetRecords(data: data)
-                return
-            default:
-                print("LOG: fail requestGetAllFriendsRecords", response)
-                NetworkAlert.show(in: self){ [weak self] in
-                    self?.requestGetAllFriendsRecords()
-                }
-                break
-            }
-        }
-        
-    }
-    
-    private func requestGetFriendCards(){
-        
-        let friendIndex = currentFriendIndex - 1
-        let friendId = friends[friendIndex].friendUserId
-        
-        FriendService.shared.getFriendRecord(id: friendId,
-                                             pageable: PageableModel(page: page),
-                                             animate: willLoadingViewAnimate){ result in
-            switch result{
-            case .success(let data):
-                print("LOG: success requestGetFriendCards", data)
-                self.processResponseGetRecords(data: data)
-                break
-            default:
-                print("LOG: fail requestGetFriendCards", result)
-                NetworkAlert.show(in: self){ [weak self] in
-                    self?.requestGetFriendCards()
-                }
-                break
-            }
-        }
-    }
-    
-    private func processResponseGetRecords(data: PageableResponseModel<RecordResponseModel>){
-        
-        hasNextPage = !data.last
-        
-        if(page == 0){
-            records = data.content
-        }else{
-            records.append(contentsOf: data.content)
-        }
-        
-        if(data.empty){
-            recordRequestIsEmpty()
-        }
-    }
-    
-    func recordRequestIsEmpty() {
-        isPaging = false
-        friendView.tableView.reloadSections(IndexSet(integer: 1), with: .none)
-    }
-    
-    
-    func requestGenerateFriendCardEmotion(reactionId: Int){
-        
-        guard let cellIndex = self.currentEmotionSelectCardIndex,
-                let reaction = Reaction(rawValue: reactionId) else { return }
-        
-        FriendService.shared.generateFriendEmotion(id: records[cellIndex].id,
-                                                   emotion: reactionId){ result in
-            switch result{
-            case .success(let data):
-                print("LOG: success requestGenerateFriendCardEmotion", data)
-                self.records[cellIndex] = data
-//                self.emoijiFloatingView?.dismiss()
-                ToastMessage.showReactionMessage(type: reaction, in: self)
-                break
-            default:
-                print("LOG: fail requestGenerateFriendCardEmotion", result)
-                NetworkAlert.show(in: self){ [weak self] in
-                    self?.requestGenerateFriendCardEmotion(reactionId: reactionId)
-                }
-                break
-            }
-        }
-    }
-    
-    private func requestHideFriendRecord(indexPath: IndexPath){
-
-        let index = dataIndexBy(indexPath)
-        let record = records[index]
-        
-        FriendService.shared.hideFriendRecord(id: record.id){ result in
-            switch result{
-            case .success:
-                print("LOG: success requestHideFriendRecord")
-                self.processResponseHideFriendRecord(indexPath: indexPath)
-                break
-            default:
-                print("LOG: fail requestHideFriendRecord", result)
-                NetworkAlert.show(in: self){ [weak self] in
-                    self?.requestHideFriendRecord(indexPath: indexPath)
-                }
-                break
-            }
-        }
-    }
-    func processResponseHideFriendRecord(indexPath: IndexPath){
-        willDelete = true
-        let recordIndex = dataIndexBy(indexPath)
-        records.remove(at: recordIndex)
-        friendView.tableView.deleteRows(at: [indexPath], with: .fade)
-        ToastMessage.showHideCompleteMessage(in: self)
-        willDelete = false
-    }
-}
-
-//MARK: - CollectionView Delegate
-extension FriendViewController: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout{
+extension FriendViewController: UICollectionViewDelegate, UICollectionViewDataSource{
     
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        friends.count + 1
+        viewModel.friends.count + 1
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-
         let cell = collectionView.dequeueReusableCell(for: indexPath, cellType: FriendCollectionViewCell.self)
-        
-        if(indexPath.row == 0){ //친구 목록 - 전체인 경우
-            cell.profileImage.image = Image.categoryInactive
-            cell.nameLabel.text = "전체"
-        }else{ //친구 목록 - 친구인 경우
-            let friend = friends[indexPath.row - 1]
-            cell.nameLabel.text = friend.friendNickName
-            cell.profileImage.kf.setImage(with: URL(string: friend.imageKey), placeholder: Image.photoDefault)
+        switch indexPath.row {
+        case 0:     return getAllFriends(cell)
+        default:    return getFriend(cell, indexPath: indexPath)
         }
-        
-        if(indexPath.row == currentFriendIndex){
-            cell.setSelectState(row: indexPath.row)
-        }
-        
-        return cell
-    }
-
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath){
-        
-        if let willDeselctCell = collectionView.cellForItem(at: [0, currentFriendIndex]) as? FriendCollectionViewCell{
-            willDeselctCell.setUnselectState(row: currentFriendIndex)
-        }
-        
-        if(currentFriendIndex == 0 && indexPath.row != 0){
-            guard let friendListCell = friendView.tableView.cellForRow(at: [0,0]) as? FriendListTableViewCell,
-                    let cell = friendListCell.collectionView.cellForItem(at: [0,0]) as? FriendCollectionViewCell else { return }
-            cell.setUnselectState(row: 0)
-        }
-        
-        guard let cell = collectionView.cellForItem(at: indexPath) as? FriendCollectionViewCell else { return }
-        cell.setSelectState(row: indexPath.row)
-        currentFriendIndex = indexPath.row
-        currentFriendIndex == 0 ? requestGetAllFriendsRecords() : requestGetFriendCards()
     }
     
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize{
-        FriendCollectionViewCell.cellSize
+    private func getAllFriends(_ cell: FriendCollectionViewCell) -> FriendCollectionViewCell{
+        cell.do{
+            $0.profileImage.image = Image.categoryInactive
+            $0.nameLabel.text = "전체"
+        }
+        return checkFriendCellSelectState(cell, index: 0)
+    }
+
+    private func getFriend(_ cell: FriendCollectionViewCell, indexPath: IndexPath) -> FriendCollectionViewCell{
+        let friend = viewModel.friends[indexPath.ofFriendData]
+        cell.do{
+            $0.nameLabel.text = friend.friendNickName
+            $0.profileImage.kf.setImage(with: URL(string: friend.imageKey), placeholder: Image.photoDefault)
+        }
+        return checkFriendCellSelectState(cell, index: indexPath.row)
+    }
+
+    private func checkFriendCellSelectState(_ cell: FriendCollectionViewCell, index: Int) -> FriendCollectionViewCell{
+        if index == selectedFriendCellIndex.value {
+            cell.setSelectState(row: index)
+        }
+        return cell
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        
+        collectionView.cellForItem(at: [0, selectedFriendCellIndex.value], cellType: FriendCollectionViewCell.self)?.do{
+            $0.setUnselectState(row: selectedFriendCellIndex.value)
+        }
+        
+        collectionView.cellForItem(at: indexPath, cellType: FriendCollectionViewCell.self)?.do{
+            $0.setSelectState(row: indexPath.row)
+        }
+        selectedFriendCellIndex.accept(indexPath.row)
     }
 }
 
-//MARK: - TableView Delegate
-extension FriendViewController: UITableViewDelegate, UITableViewDataSource, FriendRecordCellDelegate{
+extension FriendViewController: UITableViewDelegate, UITableViewDataSource{
     
     func numberOfSections(in tableView: UITableView) -> Int {
         2
     }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if(section == 0){
-            return records.count + 1
-        }else if(section == 1 && isPaging && hasNextPage){
+        if section == 0 {
+            return viewModel.records.count + COUNT_OF_NOT_RECORD_CELL
+        } else if section == 1 && isLoading && viewModel.hasNextPage {
             return 1
-        }else{
+        } else {
             return 0
         }
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-
-        if(indexPath.section == 1){
-            let cell = tableView.dequeueReusableCell(for: indexPath, cellType: LoadingTableViewCell.self)
-            cell.startLoading()
-            return cell
-        }
-        
-        if(indexPath.row == 0){
-            let cell = tableView.dequeueReusableCell(for: indexPath, cellType: FriendListTableViewCell.self)
-            cell.collectionView.delegate = self
-            cell.collectionView.dataSource = self
-            return cell
-        }
-        
-        let cell = tableView.dequeueReusableCell(for: indexPath, cellType: FriendTableViewCell.self)
-        let cardIndex = dataIndexBy(indexPath)
-        let record = records[cardIndex]
-    
-        cell.delegate = self
-        cell.bindingData(record)
-                
-        return cell
-    }
-    
-    func willShowReactionFloatingView(indexPath: IndexPath) {
-
-//        isSufficientToShowFloatingView(indexPath: indexPath){
-//
-//            currentEmotionSelectCardIndex = dataIndexBy(indexPath)
-//            emoijiFloatingView = EmojiFloatingView().then{
-//                $0.delegate = self
-//                $0.completion = {
-//                    print("LOG: emoijiFloatingView completion closure called")
-//                    self.currentEmotionSelectCardIndex = nil
-//                    self.emoijiFloatingView = nil
-//                }
-//            }
-//
-//            guard let cell = friendView.tableView.cellForRow(at: indexPath) as? FriendTableViewCell else { return }
-//            emoijiFloatingView.show(in: self, standard: cell)
-//        }
-    }
-    
-    private func isSufficientToShowFloatingView(indexPath: IndexPath, closure: () -> Void){
-        
-        let rectOfCellInTableView = friendView.tableView.rectForRow(at: indexPath) //5번째 셀의 좌표 값
-        let rectOfCellInSuperview = friendView.tableView.convert(rectOfCellInTableView, to: self.view)
-
-        let viewPosition = CGPoint(x: rectOfCellInSuperview.origin.x,
-                                   y: rectOfCellInSuperview.origin.y + rectOfCellInSuperview.height)
-        
-        if(Device.HEIGHT - viewPosition.y - self.view.safeAreaInsets.bottom >= 74){ //Device.HEIGHT - viewPosition.y - Device.tabBarHeight - self.view.safeAreaInsets.bottom
-            print("LOG: EMOJI FLOATING VIEW TEST true")
-            closure()
-        }else{
-            ToastMessage.showMakeSufficientSpaceMessage(in: self)
-            print("LOG: EMOJI FLOATING VIEW TEST falase")
-            return
+        switch indexPath.section {
+        case 0:     return getFriendTableViewCell(indexPath: indexPath)
+        default:    return getLoadingCell(indexPath: indexPath)
         }
     }
+    private func getFriendTableViewCell(indexPath: IndexPath) -> BaseTableViewCell {
+        switch indexPath.row {
+        case 0:     return getFriendsListCell(indexPath: indexPath)
+        default:    return getFriendRecordCell(indexPath: indexPath)
+        }
+    }
+    private func getFriendsListCell(indexPath: IndexPath) -> BaseTableViewCell{
+        mainView.tableView.dequeueReusableCell(for: indexPath, cellType: FriendListTableViewCell.self).then{
+            $0.collectionView.delegate = self
+            $0.collectionView.dataSource = self
+        }
+    }
+    
+    private func getFriendRecordCell(indexPath: IndexPath) -> BaseTableViewCell{
+        mainView.tableView.dequeueReusableCell(for: indexPath, cellType: FriendTableViewCell.self).then{
+            $0.delegate = self
+            $0.bindingData(viewModel.records[indexPath.ofRecordData])
+        }
+    }
+    
+    private func getLoadingCell(indexPath: IndexPath) -> BaseTableViewCell{
+        mainView.tableView.dequeueReusableCell(for: indexPath, cellType: LoadingTableViewCell.self).then{
+            $0.startLoading()
+        }
+    }
+}
+
+extension FriendViewController: FriendRecordCellDelegate{
     
     func presentReactionSheet(indexPath: IndexPath) {
-        let data = records[dataIndexBy(indexPath)].friendReactions
-        FriendReactionSheetTestViewController(reactions: data).show(in: self)
+        let reactions = viewModel.records[indexPath.ofRecordData].friendReactions
+        FriendReactionSheetViewController(reactions: reactions).show(in: self)
     }
     
     func presentEtcActionSheet(indexPath: IndexPath) {
-        let alert = UIAlertController(title: nil,
-                                      message: nil,
-                                      preferredStyle: .actionSheet)
         
-        let hideAction = UIAlertAction(title: "숨기기", style: .default){ _ in
-            alert.dismiss(animated: true)
-            self.requestHideFriendRecord(indexPath: indexPath)
-        }
+        let alert = UIAlertController(
+            title: nil,
+            message: nil,
+            preferredStyle: .actionSheet
+        )
 
-        let declarationAction = UIAlertAction(title: "신고하기", style: .default) { _ in
-            _ = LinkManager(self, .report)
-        }
-        
+        let hideAction = makeAlertHideAction(alert: alert, indexPath: indexPath)
+        let declarationAction = makeAlertDeclarationAction(alert: alert, indexPath: indexPath)
         let cancelAction = UIAlertAction(title: "취소", style: .cancel, handler: nil)
         
         alert.addAction(hideAction)
@@ -460,5 +273,75 @@ extension FriendViewController: UITableViewDelegate, UITableViewDataSource, Frie
         alert.addAction(cancelAction)
              
         present(alert, animated: true)
+    }
+    
+    private func makeAlertHideAction(alert: UIAlertController, indexPath: IndexPath) -> UIAlertAction {
+        UIAlertAction(title: "숨기기", style: .default){ _ in
+            alert.dismiss(animated: true)
+            self.viewModel.hideRecord(index: indexPath.ofRecordData)
+        }
+    }
+    
+    private func makeAlertDeclarationAction(alert: UIAlertController, indexPath: IndexPath) -> UIAlertAction {
+        UIAlertAction(title: "신고하기", style: .default) { _ in
+            LinkManager(self, .report)
+        }
+    }
+    
+    func willShowReactionFloatingView(indexPath: IndexPath) {
+        if isSufficientToShowFloatingView(indexPath: indexPath) {
+            showReactionFloatingView(indexPath: indexPath)
+        } else {
+            ToastMessage.showMakeSufficientSpaceMessage(in: self)
+        }
+    }
+    
+    private func isSufficientToShowFloatingView(indexPath: IndexPath) -> Bool {
+        
+        let rectOfCellInTableView = mainView.tableView.rectForRow(at: indexPath)
+        let rectOfCellInSuperview = mainView.tableView.convert(rectOfCellInTableView, to: view)
+
+        let viewPosition = CGPoint(
+            x: rectOfCellInSuperview.origin.x,
+            y: rectOfCellInSuperview.origin.y + rectOfCellInSuperview.height
+        )
+        
+        return Device.HEIGHT - viewPosition.y - view.safeAreaInsets.bottom >= 74
+    }
+    
+    private func showReactionFloatingView(indexPath: IndexPath){
+        if let cell = mainView.tableView.cellForRow(at: indexPath) as? FriendTableViewCell {
+            reactionFloatingView.show(in: self, standard: cell){
+                self.viewModel.registerReaction(id: $0, index: indexPath.ofRecordData)
+            }
+        }
+    }
+}
+
+extension FriendViewController{
+    
+    func scrollViewDidScroll(_ scrollView: UIScrollView) {
+        
+        let offset = scrollView.contentOffset.y
+        let contentHeight = scrollView.contentSize.height
+        let height = scrollView.frame.height
+        
+        if canPaging(offset: offset, contentHeight: contentHeight, height: height) {
+            beginPaging()
+        }
+    }
+    
+    private func canPaging(offset: CGFloat, contentHeight: CGFloat, height: CGFloat) -> Bool{
+        offset > (contentHeight - height) && !isLoading && viewModel.hasNextPage
+    }
+    
+    private func beginPaging(){
+        isLoading = true
+        DispatchQueue.main.async { [self] in
+            mainView.tableView.reloadSections(IndexSet(integer: 1), with: .none)
+        }
+        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) { [self] in
+            willPaging.onNext(())
+        }
     }
 }
